requirements.txt
django==3.0.5
djangorestframework
djongo